+++ CHANGELOG.md
@@ -9,6 +9,8 @@
  - Rebuild of Images, Sprite and SpriteAnimation initialization
  - Use isRelative on effects
  - Use Vector2 for position and size on PositionComponent
+ - Rename `resize` method on components to `onGameResize`
+ - Make `Resizable` have a `gameSize` property instead of `size`
 
 ## [next]
  - Fix spriteAsWidget deprecation message